+++ Assets/Settings.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class Settings : MonoBehaviour
 {
-    public int GridSize = 10;
+    public float GridSize = 10;
     public int CanvasWidth = 70;
     public int CanvasHeight = 50;
     // Start is called before the first frame update
